requirements.txt
opencv-python
numpy
pandas
matplotlib
scikit-learn
facenet-pytorch
pillow
tqdm
albumentations
fastapi
uvicorn
python-multipart
